Added more gestures; plays audio; added LeanTween
PlayObjects can be programmatically instantiated with support for
Tap, Drag, Press, and Release gestures. A 'highlight' object will
appear behind any currently touched object, to signal to the user
that the object is being touched.

LeanTween was added for animating sprites. Now PlayObjects will
'pulse' in place (growing a little bigger and shrinking smaller),
which is used to indicate that those objects can be interacted with.

An audio clip can be attached to a PlayObject. That clip will be
played when the object is tapped.

An asset postprocessor script was added so that all audio clips
that Unity imports will be set as 2D sounds, not 3D sounds.