Moved gesture code, fixed audio bug, prep for ROS
Refactored gesture events so that all gesture stuff is in its own
script, not in the main game controller.

Fixed the playback of a game object's audio clip on tap, so that
it only plays if it's not already playing.

Started adding framework for sending and receiving external ROS
msgs - so far, have a skeleton of a "handle received messages"
function with some of the kinds of messages we'll receive listed
(such as disable/enable touch on the screen, reload the scene,
load new objects).