Drag is smooth, colliders are polygons, minor fixes
Background loads at z=2 instead of z=0 though still not sure why
it wasn't loading at z=2 before

Drag/pan is smooth because we added Transformer2D from TouchScript
to each draggable object, which moves objects smoothly - but this
does mean enable/disable touch doesn't work (because we were checking
if touch was allowed before moving an object before, now we're not, so),
need to fix that still.

Colliders are polygons because even with extra vertices it's not super
computationally intensive for the small number of objects we have, and
makes our collision detection more accurate.