Fixed layers so touches are caught properly
When small objects were on top of larger ones, the larger ones
caught all the touch events. So with proper use of layers, and by
ordering layers in the TouchScript Touch Manager, we can say that
the layer of moveable objects should always get touches first, then
if none of them are hit, then we can check other layers.

Also updated submodule libraries.