Fixed collisions and speed of dragging objects
Collisions were triggered intermittently. Problem seems to have
been how the polygon colliders were created - was getting error
'sprite outline generation failed - could not read texture pixel
data. Did you forget to make the texture readable?' when each
image was loaded. It's only a problem when the app is deployed,
not in the unity editor. Fixed by setting all the images as
'advanced' and checking the 'read/write' box, which apparently
allows the textures to be read by scripts, not just from the
editor.

The transformer2d speed was too slow to keep up with how fast
people drag stuff on a tablet, so I upped the speed. May need
to adjust further.