Made sidekick presense configurable with in file
So now you can set sidekick=true or false in the config file, and the
sidekick will be present or absent from the basic scene as desired.
This is with a new scene called "start scene" which at the moment is the
same as the basic-scene, but theoretically will be cleaned up a bit.