Remove old check if objects on screen function
The old function for checking whether objects were still on screen after
a drag action used hard-coded values for the screen width and screen
height, which we don't want, since the screen size will change depending
on what device the game is deployed on. We also remove the old constants
holding the screen width and height values.