Update demo with touchscript components
- The demo hadn't been updated after TouchScript was last updated. Thus,
all the TouchScript scripts were missing from the game objects in the
demo scenes. They have now been added.
- The arrow in demo-scene1 and
demo-scene2 was tagged "Back" but should have been tagged "GoBack" for
it to register taps.
- Set the default flags in MainGameController to demo=true and
story=galse (i.e., make it so the demo is the default when you check out
the repo).