Fix bug where next/previous page commands worked only if touch enabled
The ChangePage function was only allowing the page in a storybook to
advance or go to the previous page when touch was enabled. However, a
common use case is changing the page during the robot's turn in the game
when touch is disabled. The check for touch being enabled has been
removed.